feat: store multiple "DBs" within a sync.Map
Allows the library to handle multiple *Database 's in concurrently
safely, but optimised towards more reads than writes. This means you can
write safely, but its optimised to only write a key once.

When "New" is used, the key stored is the DSN, primarily for backwards
compatibility. When "NewWithName" is used, the given name is the key.