feat: count number of DB's and helper method
Keeps count of number of "New" DB's with a private global atomic
integer.

Adds a warnNumDiffDBs func to check if the number of different
DBs specified is more than one, this is used to return an error, or in
this case warning, which can be ignored when setting
IgnoreUnspecifiedDBWarning.

This is implemented, since to not break the API, the global DB variable
must not be removed. Currently, the DB variable is set by the last ran
"New" or "NewWithName". Setting IgnoreUnspecifiedDBWarning will ignore
this warning and continue using the DB variable.