feat: add Use and UseC funcs
Use and UseC both are used to retrieve the *Database by a key,
atomically.

The private new additions to the Database struct, are to keep track of
whether the *Database is being accessed "intentionally".
e.g.

    mysql.New("AAAA", l.Logger)
    mysql.New("BBBB", l.Logger)

    -- Unspecified (but DB is the "BBBB" database)
    mysql.DB.Execute(XXX)

    -- Specified
    mysql.UseC("BBBB").Execute(sql, args)

The Use func returns a *Database and func to reset the "passed" value.
e.g.
    db, done := mysql.Use("BBBB")
    defer done()

    db.Execute(XXX)
    db.Execute(XXXXXX)

Whereas, the UseC func is for chaining a method.
    mysql.UseC("AAAA").Execute()

    Note: Since it returns a *Database, you could just store this DB var
    locally, however, it will return an error after the first "chained"
    method is ran. For this use case, the Use func is recommended.

Note: Currently, the implementation of a *Database being "passed
specifically" relies on a lot of atomic actions. This is useful for
development, but not preferable for production.

In which case, set:
    mysql.IgnoreUnspecifiedDBWarning = true

The compiler will check for "if" statements with this case, and
optimise, removing all related atomic actions.